backend/requirements.txt
fastapi
uvicorn
torch
torchvision
timm
ultralytics
pillow
python-multipart
minio
tensorflow
